akinaechoAPI/requirements.txt
Flask==1.1.2
joblib==1.0.1
pandas==1.2.5
python-dotenv==1.0.0
pytz==2021.1
scikit_learn==1.3.0
spotipy==2.23.0
